@@ -1,16 +1,22 @@
 name: researcher
-description:
-  Example demonstrating how to use the library merging_builder.
+description: Example demonstrating how to use the library merging_builder.
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
-dependencies:
+dependencies: null
 
 dev_dependencies:
-  build_runner: ^1.10.0
-  build_test: ^1.2.0
   researcher_builder:
     path: ../researcher_builder
-  source_gen: ^0.9.6
-  test: ^1.15.2
+  build_runner: ^1.12.2
+  build_test: ^2.0.0
+  source_gen: ^0.9.10+4
+  test: ^1.16.8
+
+dependency_overrides:
+  source_gen:
+    git:
+      url: git@github.com:dart-lang/source_gen.git
+      ref: master
+      path: source_gen
