@@ -1,19 +1,26 @@
 name: researcher_builder
-description:
-  Example demonstrating how to use the library merging_builder.
+
+description: Example demonstrating how to use the library merging_builder.
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  analyzer: ^0.39.12
-  build: ^1.3.0
-  build_config: ^0.4.2
   merging_builder:
     path: ../../
   researcher:
     path: ../researcher
-  source_gen: ^0.9.5
-  quote_buffer: ^0.1.5
+  analyzer: ^1.2.0
+  build: ^2.0.0
+  build_config: ^0.4.7
+  quote_buffer: ^0.2.2
+  source_gen: ^0.9.10+4
 
-dev_dependencies:
+dev_dependencies: null
+
+dependency_overrides:
+  source_gen:
+    git:
+      url: git@github.com:dart-lang/source_gen.git
+      ref: master
+      path: source_gen
