@@ -1,35 +1,38 @@
 name: merging_builder
 
-description:
-  MergingBuilder reads several input
-  files and writes merged output to one file.
-  StandaloneBuilder enables output to a custom
-  folder with custom output file name.
+description: MergingBuilder reads several input files and writes merged output
+  to one file. StandaloneBuilder enables output to a custom folder with custom
+  output file name.
 
-version: 0.1.6
+version: 0.2.1-nullsafety
 
 homepage: https://github.com/simphotonics/merging_builder
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  analyzer: ^0.39.15
-  build: ^1.3.0
-  dart_style: ^1.3.6
-  directed_graph: ^0.1.7
-  exception_templates: ^0.0.6
-  glob: ^1.2.0
-  meta: ^1.2.2
-  path: ^1.7.0
-  source_gen: ^0.9.6
-  lazy_evaluation: ^0.0.6+3
+  analyzer: ^1.2.0
+  build: ^2.0.0
+  dart_style: ^2.0.0
+  directed_graph: ^0.3.1
+  exception_templates: ^0.2.0
+  file: ^6.1.0
+  glob: ^2.0.0
+  lazy_memo: ^0.1.1
+  meta: ^1.3.0
+  path: ^1.8.0
+  quote_buffer: ^0.2.2
+  source_gen: ^0.9.10+4
 dependency_overrides:
-  #analyzer: 0.39.11
+  source_gen:
+    git:
+      url: git@github.com:dart-lang/source_gen.git
+      ref: master
+      path: source_gen
 
 dev_dependencies:
-  build_runner: ^1.10.0
-  build_test: ^1.2.0
-  pedantic: ^1.9.2
-  quote_buffer: ^0.1.5
-  test: ^1.15.3
+  build_runner: ^1.12.2
+  build_test: ^2.0.0
+  pedantic: ^1.11.0
+  test: ^1.16.8
