@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'package:build/build.dart';
 import 'package:exception_templates/exception_templates.dart';
+import 'package:file/local.dart';
 import 'package:glob/glob.dart';
-import 'package:lazy_evaluation/lazy_evaluation.dart';
-import 'package:meta/meta.dart';
+import 'package:lazy_memo/lazy_memo.dart';
 import 'package:path/path.dart' as path;
 import 'package:source_gen/source_gen.dart' show Generator, LibraryReader;
 
@@ -36,17 +36,17 @@ class StandaloneBuilder<S extends SyntheticInput> extends SyntheticBuilder<S> {
   StandaloneBuilder({
     String inputFiles = 'lib/*.dart',
     this.outputFiles = 'lib/standalone_(*).dart',
-    @required this.generator,
+    required this.generator,
     String header = '',
     String footer = '',
-    Formatter formatOutput,
+    Formatter? formatOutput,
     String root = '',
   })  : root = root.trim(),
         super(
             inputFiles: inputFiles,
             header: header,
             footer: footer,
-            formatOutput: formatOutput) {
+            formatter: formatOutput) {
     _resolvedOutputFiles = Lazy(_outputPaths);
   }
 
@@ -62,7 +62,7 @@ class StandaloneBuilder<S extends SyntheticInput> extends SyntheticBuilder<S> {
 
   /// Lazily computes the output file names by replacing the
   /// placeholder `(*)` in [outputFiles] with the input file basename.
-  Lazy<List<String>> _resolvedOutputFiles;
+  late final Lazy<List<String>> _resolvedOutputFiles;
 
   /// The root directory of the package the build is applied to.
   /// This variable does not need to be set if the build command is initiated
@@ -76,14 +76,14 @@ class StandaloneBuilder<S extends SyntheticInput> extends SyntheticBuilder<S> {
   /// type parameter [S].
   @override
   Map<String, List<String>> get buildExtensions => {
-        syntheticInput.value: _resolvedOutputFiles.value,
+        syntheticInput.value: _resolvedOutputFiles(),
       };
 
   @override
   FutureOr<void> build(BuildStep buildStep) async {
     final libAssetIds = await libraryAssetIds(buildStep);
     // Accessing libraries.
-    for (final libAssetId in libAssetIds ?? <AssetId>[]) {
+    for (final libAssetId in libAssetIds) {
       final library = LibraryReader(
         await buildStep.resolver.libraryFor(libAssetId),
       );
@@ -96,7 +96,7 @@ class StandaloneBuilder<S extends SyntheticInput> extends SyntheticBuilder<S> {
           buildStep.inputId.package,
           _outputFile(libAssetId),
         ),
-        arrangeContent(await generator.generate(library, buildStep),
+        arrangeContent((await generator.generate(library, buildStep)) ?? '',
             generatedBy: 'Generated by ${generator.runtimeType}. '),
       );
     }
@@ -114,7 +114,9 @@ class StandaloneBuilder<S extends SyntheticInput> extends SyntheticBuilder<S> {
     SyntheticInput.validatePath<S>(inputFiles);
     SyntheticInput.validatePath<S>(outputFiles);
     final resolvedInputFiles = Glob(inputFiles);
-    for (final inputEntity in resolvedInputFiles.listSync(root: root)) {
+    final fileSystem = LocalFileSystem();
+    for (final inputEntity
+        in resolvedInputFiles.listFileSystemSync(fileSystem, root: root)) {
       final basename = path.basenameWithoutExtension(inputEntity.path);
       final outputFileName = outputFiles.replaceAll(
         RegExp(r'\(\*\)'),
