@@ -3,22 +3,35 @@ import 'package:dart_style/dart_style.dart';
 import 'package:directed_graph/directed_graph.dart';
 import 'package:exception_templates/exception_templates.dart';
 import 'package:glob/glob.dart';
-import 'package:meta/meta.dart';
 
 import 'formatter.dart';
 import 'synthetic_input.dart';
+
+/// The default formatting function which simply returns `input`.
+String defaultFormatter(String input) => input;
 
 /// Base class of a builder that uses synthetic input.
 ///
 /// For more information about synthetic input see:
 /// [Writing an Aggregate Builder](https://github.com/dart-lang/build/blob/master/docs/writing_an_aggregate_builder.md#writing-the-builder-using-a-synthetic-input).
 abstract class SyntheticBuilder<S extends SyntheticInput> implements Builder {
+  /// Super constructor of an object of type `SyntheticBuilder`.
+  /// * `inputFiles`: Path to the input files relative to the
+  /// package root directory. Glob-style syntax is
+  /// allowed for example: `lib/*.dart`.
+  /// * `header`: `String` that will be inserted at the top of the
+  /// generated file below the 'DO NOT EDIT' warning message.
+  /// * `footer`: String that will be inserted at the very bottom of the
+  /// generated file.
+  /// * `formatter`: A function with signature `String Function(String input)`
+  /// that is used to format the generated source code.
+  /// The default formatter is: `DartFormatter().format`.
   SyntheticBuilder({
-    @required this.inputFiles,
-    this.header,
-    this.footer,
-    Formatter formatOutput,
-  })  : formatter = formatOutput ?? DartFormatter().format,
+    required this.inputFiles,
+    this.header = '',
+    this.footer = '',
+    Formatter? formatter,
+  })  : formatter = formatter ?? DartFormatter().format,
         syntheticInput = SyntheticInput.instance<S>();
 
   /// Input files. Specify the complete path relative to the
@@ -85,58 +98,56 @@ abstract class SyntheticBuilder<S extends SyntheticInput> implements Builder {
     return result;
   }
 
-  /// Returns a list of library asset ids ordered in reverse topological
+  /// Returns a ordered set of library asset ids ordered in reverse topological
   /// dependency order.
   /// * If a file B includes a file A, then A will be appear
   /// before B.
   /// * Throws [ErrorOf] if a dependency cycle is detected.
-  Future<List<AssetId>> orderedLibraryAssetIds(BuildStep buildStep) async {
-    final assetGraph = DirectedGraph<AssetId>({},
-        comparator: ((v1, v2) => -v1.data.compareTo(v2.data)));
+  Future<Set<AssetId>> orderedLibraryAssetIds(BuildStep buildStep) async {
+    final assetGraph =
+        DirectedGraph<AssetId>({}, comparator: ((v1, v2) => -v1.compareTo(v2)));
 
     // An assetId map of all input libraries with the uri as key.
-    final assetMap = <Uri, Vertex<AssetId>>{};
+    final assetMap = <Uri, AssetId>{};
 
     // Access libraries
     await for (final input in buildStep.findAssets(Glob(inputFiles))) {
       // Check if input file is a library.
       if (await buildStep.resolver.isLibrary(input)) {
-        assetMap[input.uri] = Vertex<AssetId>(input);
-        assetGraph.addEdges(assetMap[input.uri], []);
+        assetMap[input.uri] = input;
+        assetGraph.addEdges(assetMap[input.uri]!, {});
       }
     }
 
-    for (final Vertex<AssetId> vertex in assetGraph) {
-      final connectedVertices = <Vertex<AssetId>>[];
+    for (final assetId in assetGraph) {
+      final importedAsseIds = <AssetId>{};
 
       // Read library.
-      final library = await buildStep.resolver.libraryFor(vertex.data);
+      final library = await buildStep.resolver.libraryFor(assetId);
       // Get dependencies
       for (final import in library.imports) {
         if (import.uri == null) continue;
-        final uri = Uri.parse(import.uri);
+        final uri = Uri.parse(import.uri!);
         // Skip if uri scheme is not "package" or "asset".
         if (uri.scheme == 'package' ||
             uri.scheme == 'asset' ||
             uri.scheme == '') {
           // Normalise uri to handle relative and package import directives.
-          final importedAssetId =
-              AssetId.resolve(import.uri, from: vertex.data);
-          if (importedAssetId == null) continue;
+          final importedAssetId = AssetId.resolve(uri, from: assetId);
           // Add vertex matching import directive.
           if (assetMap[importedAssetId.uri] != null) {
-            connectedVertices.add(assetMap[importedAssetId.uri]);
+            importedAsseIds.add(assetMap[importedAssetId.uri]!);
           }
         }
       }
-      assetGraph.addEdges(vertex, connectedVertices);
+      assetGraph.addEdges(assetId, importedAsseIds);
     }
 
-    final topologicalOrdering = assetGraph.topologicalOrdering;
+    final topologicalOrdering = assetGraph.sortedTopologicalOrdering;
 
     if (topologicalOrdering == null) {
       // Find the first cycle
-      final cycle = assetGraph.cycle.map<AssetId>((vertex) => vertex.data);
+      final cycle = assetGraph.cycle;
 
       throw ErrorOf<SyntheticBuilder>(
           message: 'Circular dependency detected.',
@@ -146,8 +157,6 @@ abstract class SyntheticBuilder<S extends SyntheticInput> implements Builder {
     }
 
     // Return reversed topological ordering of asset ids.
-    return assetGraph.topologicalOrdering.reversed
-        .map<AssetId>((vertex) => vertex.data)
-        .toList();
+    return topologicalOrdering;
   }
 }
