@@ -1,12 +1,12 @@
 // GENERATED CODE. DO NOT MODIFY. Generated by AddNamesGenerator.
 
 // Header specified in build.yaml.
-final name0 = ['Thomas', 'Mayor'];
-final name1 = ['Philip', 'Martens'];
+final name0 = ['Philip', 'Martens'];
+final name1 = ['Thomas', 'Mayor'];
 
 final List<List<String>> names = [
-  ['Thomas', 'Mayor'],
   ['Philip', 'Martens'],
+  ['Thomas', 'Mayor'],
 ];
 
 // Footer specified in build.yaml.
